% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteo_access_token.R
\name{criteo_access_token}
\alias{criteo_access_token}
\title{Getting Criteo Access Token}
\usage{
criteo_access_token(tok = "criteo_token.RDS")
}
\arguments{
\item{tok}{file or object with authentification data. Object format:

You should use this fields (recommended list):
\itemize{
  \item token = <access token [class: chr]>,
  \item client_id = <Login (Client ID) Criteo REST API user [class: chr]>,
  \item client_secret = <Client Secret Criteo REST API user [class: chr]>,
  \item create_time = <Creating time [class: POSIXct]>
}}
}
\value{
String "Bearer <access token>" for using in other Criteo queries.
}
\description{
Getting Criteo valid token object. Updates token if then expired and returning
 valid access token.
 Loads token to virtual environment in function, that not junking global env.
}
