% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteo_advertisers.R
\name{criteo_advertisers}
\alias{criteo_advertisers}
\title{Getting advertisers in Criteo account}
\usage{
criteo_advertisers(tok = criteo_access_token(),
                   to_query_output = T)
}
\arguments{
\item{tok}{Criteo access token. By default is call and refresh token automatically.}

\item{to_query_output}{If TRUE, return comma separated advertiser id's;
if not, return advertisers data frame.}
}
\value{
By default, string of all Criteo advertisers (for using in other functions).
}
\description{
Getting advertisers in Criteo account
}
\examples{
# Base variant
criteo_advertisers()

# If you want use token from environment
criteo_advertisers(tok = token)

# If you want to get data frame
criteo_advertisers(to_query_output = F)
}
