% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteo_auth.R
\name{criteo_auth}
\alias{criteo_auth}
\title{Criteo REST API Auth}
\usage{
criteo_auth(client_id = NULL, client_secret = NULL, save_to_env = F)
}
\arguments{
\item{client_id}{Login (Client ID) of Criteo REST API user.}

\item{client_secret}{Client Secret of Criteo REST API user.}

\item{save_to_env}{by default, False. If TRUE, saving auth data into global environment.}
}
\value{
By default, file "criteo_tokes.RDS" into folder with script and success message.

This is list with 4 parameters:
\itemize{
  \item token: returned token with 300 seconds lifetime.
  \item client_id: your client id from Criteo REST API. It need for future calls.
  \item client_secret: your client secret from Criteo REST API. It need for future calls.
  \item create_time: created time for auto-refresh.
}
}
\description{
Create token object and save it into .RDS file. Return object, if needed.
You need to create REST API user in Criteo Marketing account.
}
\examples{
# Authentification in first package use
criteo_auth("mapi-41jf7549-fa1r-90af-pt54-o5jfhan91nty",
            "D1s1by_X)Dgasd;g!*") # it is not real auth data

# If you want also save all data to global environment.
criteo_auth("mapi-41jf7549-fa1r-90af-pt54-o5jfhan91nty",
            "D1s1by_X)Dgasd;g!*",
            save_to_env = T)
}
