% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteo_budget_info.R
\name{criteo_budget_info}
\alias{criteo_budget_info}
\title{Getting information about Criteo budgets}
\usage{
criteo_budget_info(tok = criteo_access_token(),
  return_columns = c("budgetName", "type", "totalAmount",
  "remainingBudget"), active_budgets_only = T)
}
\arguments{
\item{tok}{Criteo access token. By default is call and refresh token automatically.}

\item{return_columns}{vector of column names, by default:
\itemize{
  \item budgetName
  \item type
  \item totalAmount
  \item remainingBudget
}
If you want see all columns, use "all" as parameter. All columns:
\itemize{
  \item advertiserId
  \item budgetId
  \item budgetName
  \item type
  \item totalAmount
  \item remainingBudget
  \item remainingBudgetUpdated
  \item active
}}

\item{active_budgets_only}{by default, TRUE.}
}
\value{
By default, data frame with active budgets and only 4 columns (this is not all columns).
}
\description{
Getting information about Criteo budgets
}
\examples{
# Get all active budgets
criteo_budget_info()

# Get all exist budgets
criteo_budget_info(active_budgets_only = F)

Get all info about all budgets
criteo_budget_info(return_columns = "all", active_budgets_only = F)
}
