% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteo_campaigns.R
\name{criteo_campaigns}
\alias{criteo_campaigns}
\title{Getting information about Criteo campaigns}
\usage{
criteo_campaigns(tok = criteo_access_token(),
                 active_campaigns_only = T)
}
\arguments{
\item{tok}{Criteo access token. Default is call and refresh token automatically.}

\item{active_campaigns_only:}{By default TRUE.}
}
\value{
By default, data frame with active companies.
}
\description{
Getting information about Criteo campaigns
}
\examples{
# If you want to get all campaigns in account
criteo_campaigns(active_campaigns_only = F)
}
